Obfuscate passwords in serialized credentials collection
Veil now encrypts the "credentials" section of the JSON file when
writing it to disk. The key and other encryption parameters are stored
in the new "cipher" key. This obfuscation helps users who have
requirements around plaintext passwords.

We currently encrypt using AES-256-CBC. We could switch this over to a
more modern mode such as AES-256-GCM; however, since this is currently
just for compliance purposes, we went with something we assumed would
be widely supported.

Exciting, unencrypted secrets files will be encrypted when the file is
saved, providing a migration path for Chef Server installations.

Signed-off-by: Stephan Renatus <srenatus@chef.io>