[chef_secrets] Use helper from veil to export secrets to erlang
Our custom crypto module that is in use in FIPS-enabled builds does
not have support for any block ciphers. Thus, instead of
re-implementing the crypto on the erlang side, this PR adds a new
helper command that we shell out to in order to get the secrets in a
plain JSON format that requires no parsing.

Some alternatives to this include:

- Implementing the block_encrypt and block_decrypt functions required
  in our FIPS-crypto erlang module.

- Using a simple gzip+base64 for now

Signed-off-by: Steven Danna <steve@chef.io>