Hack my way around to getting the filehandle for a file.
Our tests didn't survive the OTP-21 improvements.

We relied on some non-portable behavior to create a file, and get the
filehandle to pass to chef-secrets. This broke.

First of all. OTP-21 rewrote raw files to use a nif. This has the
effect of burying the file descriptor deep in an opaque reference, and
I've not spotted a sane way to extract that info. So we dive into the
proc filesystem to find the filehandle corresponding to the file we
just opened. This method only works on linux, and I've not found an
equivalent for MacOS.

Second, we technically were leaking the filehandle; we never used it
after open, and OTP-21's GC found and cleaned it up too quickly for
the test.

Signed-off-by: Mark Anderson <mark@chef.io>